+++ pynopegl-utils/pynopegl_utils/tests/cmp.py
@@ -91,7 +91,7 @@ class CompareSceneBase(CompareBase):
         self._exercise_serialization = exercise_serialization
         self._exercise_dot = exercise_dot
         self._samples = samples
-        self._hud = 0
+        self._hud = False
         self._hud_export_filename = None
 
     def render_frames(self):
@@ -109,15 +109,17 @@ class CompareSceneBase(CompareBase):
         capture_buffer = bytearray(width * height * 4)
         ctx = ngl.Context()
         ret = ctx.configure(
-            offscreen=1,
-            width=width,
-            height=height,
-            backend=get_backend(backend) if backend else ngl.BACKEND_AUTO,
-            samples=self._samples,
-            clear_color=self._clear_color,
-            capture_buffer=capture_buffer,
-            hud=self._hud,
-            hud_export_filename=self._hud_export_filename,
+            ngl.Config(
+                offscreen=True,
+                width=width,
+                height=height,
+                backend=get_backend(backend) if backend else ngl.Backend.AUTO,
+                samples=self._samples,
+                clear_color=self._clear_color,
+                capture_buffer=capture_buffer,
+                hud=self._hud,
+                hud_export_filename=self._hud_export_filename,
+            )
         )
         assert ret == 0
         timescale = duration / float(self._nb_keyframes)