+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 ## [Unreleased]
 ### Changed
 - CSV export in the HUD now always prints floats in C locale instead of quoted
+- `pynopegl.Context.configure()` now takes a `Config` as argument
 
 ## [2023.1] [libnopegl 0.8.0] - 2023-04-07
 ### Changed